requirements.txt
Flask
PyMuPDF
python-docx